Controller & Collision updates
Updated the PlayerMover script with some components from PlayerController for the "real" game.  Updated EnemyController & PlayerEnemyCollision to check for which scene the game is in, then use either the PlayerController or PlayerMover object, as appropriate.  Need to update PlayerDeath to account for the correct scene as well, so the correct Health object is referenced.